requirements.txt
aiofiles==0.6.0
appdirs==1.4.4
black==20.8b1
click==7.1.2
cycler==0.10.0
decorator==4.4.2
easyocr==1.2.1
fastapi==0.62.0
h11==0.11.0
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
mypy-extensions==0.4.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
pathspec==0.8.1
Pillow==8.0.1
pyaml==20.4.0
pydantic==1.7.3
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
python-multipart==0.0.5
PyWavelets==1.1.1
PyYAML==5.3.1
regex==2020.11.13
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
starlette==0.13.6
tifffile==2020.12.8
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
uvicorn==0.13.1
